move input command to a shared file
