+++ management/commands/boolean_input.py
@@ -1,0 +1,7 @@
+def boolean_input(question, default=None):
+    result = input("%s " % question)
+    if not result and default is not None:
+        return default
+    while len(result) < 1 or result[0].lower() not in "yn":
+        result = input("Please answer yes or no: ")
+    return result[0].lower() == "y"
