+++ management/commands/delete_journal.py
@@ -5,13 +5,7 @@ from core.models import File
 import os
 from django.conf import settings
 
-def boolean_input(question, default=None):
-    result = input("%s " % question)
-    if not result and default is not None:
-        return default
-    while len(result) < 1 or result[0].lower() not in "yn":
-        result = input("Please answer yes or no: ")
-    return result[0].lower() == "y"
+from .boolean_input import *
 
 class Command(BaseCommand):
     """Fully deletes a journal including all associated files and objects"""
